ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
numpy
pandas
pdfplumber==0.10.3
python-docx==1.0.1
openpyxl==3.1.2
beautifulsoup4==4.12.2
markdown==3.5
python-jose==3.3.0
loguru==0.7.2
tenacity==8.2.3
scikit-learn 
faiss-cpu
openai
